feat: move ErrorResponse into error sub-dir
