+++ src/JsonService.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-import { ErrorResponse } from "./ErrorResponse";
+import { ErrorResponse } from "./errors";
 import { Logger } from "./utils";
 
 /**
